+++ src/bract/core/keydef.clj
@@ -61,6 +61,9 @@
   ctx-runtime-info   [:bract.core/runtime-info  fn-coll? "Runtime-info functions [(fn []) ..]" {:default []}]
   ctx-alive-tstamp   [:bract.core/alive-tstamp   ifn?    "Derefable (fn []): alive timestamp in milliseconds"
                       {:default (util/alive-millis)}]
+  ctx-jvm-exit-code  [:bract.core/jvm-exit-code  (some-fn (every-pred integer? (some-fn pos? zero?))
+                                                   nil?) "JVM exit code (int) - zero or more" {:parser kputil/any->int
+                                                                                               :default nil}]
   *ctx-shutdown-flag [:bract.core/*shutdown-flag volatile? "Volatile: Shutdown begun?" {:default (volatile! false)}]
   ctx-shutdown-hooks [:bract.core/shutdown-hooks vector? "Added shutdown hook threads" {:default []}])